Connect directly to backend rather than go through gateway
At the moment, the gateway won't start without a working URL
for the scheduler backend. To avoid having to run the schduler
backend, we can point the frontend directly to the core backend.
To do this, we assign the /graphql path to the core backend instead of
the gateway when running the E2E tests.

Part of UserOfficeProject/stfc-user-office-project#221