refactor: Use duo-logger instead of console everywhere in the main app
Using duo-logger for all messages ensures that all log output is formatted
in the same way, allowing us to parse and monitor it with tools like Elasticsearch.

It also makes it easier to see where messages are being logged, and to replace
the default logger if necessary.

Part of UserOfficeProject/stfc-user-office-project#231