Set Compose project dir to current working dir
Without this, the project directory is set as the directory
containing the first Compose file passed to docker-compose.

Now that the first Compose file is in the user-office-backend
directory, the mount paths for all apps would be wrong
without setting the project dir to the working dir.

Part of UserOfficeProject/stfc-user-office-project#221