Add validation to file question
Adds file upload validation. The validation looks ateach file's signature to check that the file uploaded is one of the accepted file types.

Validation for an individual file passes if the file can be read and determined to be one of the accepted types, or if there is a problem retrieving the file to validate it.

Validation for an individual file fails if the file can be read and is not one of the accepted types, or if no types can be analysed from the file.

For the entire question to pass validation, all files within the question must pass the invididual file validation.

Part of UserOfficeProject/stfc-user-office-project#226