feat: Auto-assign STFC instrument sciensts to instruments
When using STFC external authentication, users in one of the
instrument scientists STFC roles are now automatically assigned
to the instruments for their facility.

This feature requires the instruments to be in the DB when the user
logs in, otherwise no instruments will be assigned. The instruments
are identified by name, to make it easy to recreate then if they
were deleted.

The auto-assign code also automatically unassigns instruments
if the user is no longer in the instrument scientist group
or if the instrument was removed.

Part of https://github.com/UserOfficeProject/user-office-project-issue-tracker/issues/1