feat: Show proposals for assigned calls in instrument scientist table
The instrument scientist proposal table now shows all proposals for all calls
assigned to the scientist's instruments. This is in addition to the proposals
assigned directly to the scietist's instruments, which is what the table showed
before this change.

The main reason for this is to automatically show all proposals received
for a given instrument without requiring a user officer to assign each proposal
to the instrument.

Part of https://github.com/UserOfficeProject/user-office-project-issue-tracker/issues/1