fix: Only show proposals for instrument's call in STFC mode
The STFC setup goes back to using a constom ProposalDataSource
so that it can show all proposals for calls assigned to an instrument,
while the ESS setup keeps the previous behavior of only showing
proposals directly assigned to an instrument.

This commit also updates the ProposalView record class
to match the updated database view, which I had forgotten
when I regenerated the view. The proposal instrument is now either
the instrument assigned to the proposal (if present)
or the instrument assigned to its call.

Part of https://github.com/UserOfficeProject/user-office-project-issue-tracker/issues/1