Remove duplicate proposals in instrument scientist view
This commit reverts the recent changes to the proposal table view
which produced duplicate records when the proposal was assigned to
a call with multiple instruments.

STFC still needs to show proposals for a call's instrument
rather than proposals directly assigned to an instrument,
so the logic for doing this has been moved to the STFC ProposalDataSource.
This will likely slow down the STFC instrument scientist page,
but avoid the complexity of making the DB view change work
for both facilities.

Part of https://github.com/UserOfficeProject/user-office-project-issue-tracker/issues/609