performance: Regenerate UOWS client with v3.0.0 of the generator
Regenerates the UOWS client using the latest generator version.
This includes changes which reduce calls made to the UOWS (from
https://github.com/UserOfficeProject/user-office-lib/pull/131 and
https://github.com/UserOfficeProject/user-office-lib/pull/134).

As part of this, Typescript was automatically updated to its latest version,
as the uows-client-generator package now requires a later version of Typescript
than the one previously used in this app. The Eslint plugins have been updated
to match the new Typescript version.

Part of https://github.com/UserOfficeProject/user-office-project-issue-tracker/issues/580