fix: Correct checking if user is a SEP member
Wait until the DB call has returned before deciding if the user
is a SEP member. Without this, the check would always pass as the
function returns a Promise, which is always defined even if the result
was a boolean false.

This was spotted by the new version of TypeScript, which now raises
a linter error if these kind of checks are used on Promises.