feat: Speed up STFC setup by caching UOWS results
The backend now caches most results from the UOWS.
This includes validity of user tokens, person details,
and roles. The data is stored in an LRU cache for a short time,
so that any requests for the same data will not need a trip
to the UOWS.

The "getUsers" function in the STFC user datasource has been left
"uncached", as it's called very frequently as the user types
surnames in a dialog box. I suspect that trying to cache results
there would only lead to a large amount of useless results
flooding the cache.

To help with evaluating the effectiveness of the caches,
I've set them to log usage statistics every 5 minutes.
This can be changes if people feel it would crowd the logs too much.

As part of this, I've implmented the cache myself. The main reason
was that most of the caches I found didn't quite fit our use case
in here, either because they couldn't auto-remove results or they
needed separate instances of standalone cache DBs. I can use an
existing one if people prefer.

Part of https://github.com/UserOfficeProject/user-office-project-issue-tracker/issues/580