Add manual mock for LRUCache in unit tests
Clling the 'enableStatsLogging' function in the unit tests
can still sometimes prevent Jest from exiting, as it does not
try to stop the timer started by that function.

The manual mock can now be used instead of the real cache when testing
units which rely on the cache. This solves the problem with
timers and may prevent confusion in cases where we're trying to test
units but the cache still has results from a previous test.

Part of https://github.com/UserOfficeProject/user-office-project-issue-tracker/issues/580