fix: Improve handling of GraphQL errors
The frontend now checks if the error has an empty response,
which should handle the case where the server didn't respond in time.
This seems to sometimes happen when the application is running
unser high load.

GraphQL errors are now logged to both the client and server logs,
with a copy of the query which caused the error.

The request and error handling functions are now correctly typed
to ensure that values are being treated correctly. The 'request'
function now rethrows errors rather than returning them to its caller,
which ensures the caller always gets a value of the expected type.

Part of https://github.com/UserOfficeProject/user-office-project-issue-tracker/issues/18