feat(config): add support for creating an application event bus in dependencyConfigE2E, dependencyConfigSTFC, and dependencyConfigTest files
The application event bus is created using the `createApplicationEventBus` function from the `../events` module. This event bus will be used for handling and dispatching events within the application.

This change ensures that the event bus is properly configured and available for use in the E2E, STFC, and Test environments.