PDF Preview Page done.
Changed env variable

Fixed the bug of proposal SEP assignment, when Instrument picker is answered and the Instrument is assigned to the Proposal

Added new column for storing dummy json data for Proposal.

Writing the url bacjk.

Created endpoint for PDF Preview and calling it.

Code cleanup and optimization

Added pdf template data