feat: Add support for initial user officer role assignment
This commit adds support for assigning the USER_OFFICER role to the initial user officer based on the value of the `INITIAL_USER_OFFICER_EMAIL` environment variable. If the email matches the value of `INITIAL_USER_OFFICER_EMAIL`, the USER_OFFICER role is assigned. Otherwise, the USER role is assigned. If `INITIAL_USER_OFFICER_EMAIL` is not set or if there is no email provided, the USER role is assigned.